Update/improve more mixin class names
