+++ fabric-blockrenderlayer-v1/src/main/resources/fabric-blockrenderlayer-v1.mixins.json
@@ -3,7 +3,7 @@
   "package": "net.fabricmc.fabric.mixin.blockrenderlayer",
   "compatibilityLevel": "JAVA_8",
   "client": [
-    "MixinBlockRenderLayer"
+    "MixinRenderLayers"
   ],
   "injectors": {
     "defaultRequire": 1
