+++ fabric-commands-v0/src/main/resources/fabric-commands-v0.mixins.json
@@ -6,7 +6,7 @@
     "MixinMinecraftDedicatedServer"
   ],
   "client": [
-    "MixinCommandManagerIntegrated"
+    "MixinCommandManager"
   ],
   "injectors": {
     "defaultRequire": 1
