+++ fabric-dimensions-v1/src/main/resources/fabric-dimensions-v1.mixins.json
@@ -7,7 +7,7 @@
     "MixinEntity",
     "MixinPortalForcer",
     "MixinWorld",
-    "idremap.MixinDimensionRawIndexFix",
+    "idremap.MixinDimensionType",
     "idremap.MixinLevelProperties",
     "idremap.MixinLevelStorage",
     "idremap.MixinPlayerManager",
