+++ fabric-keybindings-v0/src/main/resources/fabric-keybindings-v0.mixins.json
@@ -5,7 +5,7 @@
   "client": [
     "MixinGameOptions",
     "MixinKeyBinding",
-    "KeyCodeAccessor"
+    "KeyBindingAccessor"
   ],
   "injectors": {
     "defaultRequire": 1
