+++ fabric-registry-sync-v0/src/main/resources/fabric-registry-sync-v0.mixins.json
@@ -5,7 +5,7 @@
   "mixins": [
     "MixinBootstrap",
     "MixinIdList",
-    "MixinIdRegistry",
+    "MixinSimpleRegistry",
     "MixinPlayerManager",
     "MixinWorldSaveHandler"
   ],
