+++ fabric-renderer-indigo/src/main/java/net/fabricmc/fabric/mixin/client/indigo/renderer/MixinChunkData.java
@@ -27,7 +27,7 @@ import net.minecraft.client.render.chunk.ChunkBuilder.ChunkData;
 import net.fabricmc.fabric.impl.client.indigo.renderer.accessor.AccessChunkRendererData;
 
 @Mixin(ChunkData.class)
-public class MixinChunkRenderData implements AccessChunkRendererData {
+public class MixinChunkData implements AccessChunkRendererData {
 	@Shadow
 	private Set<RenderLayer> initializedLayers;
 	@Shadow
