+++ fabric-rendering-data-attachment-v1/src/main/resources/fabric-rendering-data-attachment-v1.mixins.json
@@ -4,7 +4,7 @@
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "MixinBlockEntity",
-    "MixinViewableWorld"
+    "MixinWorldView"
   ],
   "client": [
     "client.MixinChunkRendererRegion"
