+++ fabric-rendering-v1/src/main/java/net/fabricmc/fabric/mixin/client/rendering/MixinBlockColors.java
@@ -23,28 +23,27 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+import net.minecraft.block.Block;
+import net.minecraft.client.color.block.BlockColorProvider;
 import net.minecraft.client.color.block.BlockColors;
-import net.minecraft.client.color.item.ItemColorProvider;
-import net.minecraft.client.color.item.ItemColors;
-import net.minecraft.item.ItemConvertible;
 import net.minecraft.util.IdList;
 import net.minecraft.util.registry.Registry;
 
 import net.fabricmc.fabric.impl.client.rendering.ColorProviderRegistryImpl;
 
-@Mixin(ItemColors.class)
-public class MixinItemColorMap implements ColorProviderRegistryImpl.ColorMapperHolder<ItemConvertible, ItemColorProvider> {
+@Mixin(BlockColors.class)
+public class MixinBlockColors implements ColorProviderRegistryImpl.ColorMapperHolder<Block, BlockColorProvider> {
 	@Shadow
 	@Final
-	private IdList<ItemColorProvider> providers;
+	private IdList<BlockColorProvider> providers;
 
 	@Inject(method = "create", at = @At("RETURN"))
-	private static void create(BlockColors blockMap, CallbackInfoReturnable<ItemColors> info) {
-		ColorProviderRegistryImpl.ITEM.initialize(info.getReturnValue());
+	private static void create(CallbackInfoReturnable<BlockColors> info) {
+		ColorProviderRegistryImpl.BLOCK.initialize(info.getReturnValue());
 	}
 
 	@Override
-	public ItemColorProvider get(ItemConvertible item) {
-		return providers.get(Registry.ITEM.getRawId(item.asItem()));
+	public BlockColorProvider get(Block block) {
+		return providers.get(Registry.BLOCK.getRawId(block));
 	}
 }
