+++ fabric-rendering-v1/src/main/resources/fabric-rendering-v1.mixins.json
@@ -3,9 +3,9 @@
   "package": "net.fabricmc.fabric.mixin.client.rendering",
   "compatibilityLevel": "JAVA_8",
   "client": [
-    "MixinBlockColorMap",
+    "MixinBlockColors",
     "MixinInGameHud",
-    "MixinItemColorMap",
+    "MixinItemColors",
     "MixinWorldRenderer"
   ],
   "injectors": {
