+++ fabric-resource-loader-v0/src/main/java/net/fabricmc/fabric/mixin/resource/loader/MixinMinecraftClient.java
@@ -41,7 +41,7 @@ import net.minecraft.resource.ResourceType;
 import net.fabricmc.fabric.impl.resource.loader.ModResourcePackUtil;
 
 @Mixin(MinecraftClient.class)
-public class MixinMinecraftGame {
+public class MixinMinecraftClient {
 	@Shadow
 	private ReloadableResourceManager resourceManager;
